# Optimized bstream reader used by XORChunk iterator