I spent the last couple of days investigating query performances and I've noticed that, for several queries we see in production, a significant amount of time is spent iterating series and sample. In this analysis, I've noticed most of the time spent iterating samples is in the `XORChunk` iterator and further benchmarks showed that the `bstream` reader introduces a significant bottleneck.

I've tried different approaches to optimise the `bstream` reader and in this PR I'm proposing the optimisation I've found having the least downsides. The optimisation is split into two parts:
1. Work with a `uint64` buffer in order to optimise bit-wise operations (assuming most production architectures are 64 bit).
2. Introduce `Fast()` version of read bit/bits which get inlined by the compiler. These functions must be leafs and super small, reason why may return an error in case the current buffer  doesn't contain enough bits to satify the request. In case a `Fast()` function returns error, it must be retried using the normal (but slower) one.

The following show the result of a couple of benchmarks. I've measured similar query reduction times running benchmarks in Cortex too.

## BenchmarkXORIterator

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkXORIterator-4     33.0          22.7          -31.21%

benchmark                  old allocs     new allocs     delta
BenchmarkXORIterator-4     0              0              +0.00%

benchmark                  old bytes     new bytes     delta
BenchmarkXORIterator-4     1             1             +0.00%
```

## BenchmarkRangeQuery

```
benchmark                                                                                                             old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_one,steps=1-4                                                                              15400          14621          -5.06%
BenchmarkRangeQuery/expr=a_one,steps=10-4                                                                             15519          14847          -4.33%
BenchmarkRangeQuery/expr=a_one,steps=100-4                                                                            23341          22079          -5.41%
BenchmarkRangeQuery/expr=a_one,steps=1000-4                                                                           86703          76500          -11.77%
BenchmarkRangeQuery/expr=a_ten,steps=1-4                                                                              74669          65145          -12.75%
BenchmarkRangeQuery/expr=a_ten,steps=10-4                                                                             74867          66844          -10.72%
BenchmarkRangeQuery/expr=a_ten,steps=100-4                                                                            155213         135659         -12.60%
BenchmarkRangeQuery/expr=a_ten,steps=1000-4                                                                           806163         677584         -15.95%
BenchmarkRangeQuery/expr=a_hundred,steps=1-4                                                                          669575         591655         -11.64%
BenchmarkRangeQuery/expr=a_hundred,steps=10-4                                                                         700674         611782         -12.69%
BenchmarkRangeQuery/expr=a_hundred,steps=100-4                                                                        1486458        1288216        -13.34%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-4                                                                       7768523        7145822        -8.02%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-4                                                                    20325          22230          +9.37%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-4                                                                   20893          20294          -2.87%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-4                                                                  37490          35440          -5.47%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-4                                                                 180701         171688         -4.99%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-4                                                                    86529          103029         +19.07%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-4                                                                   95056          86269          -9.24%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-4                                                                  262109         238824         -8.88%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-4                                                                 1676112        1558172        -7.04%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-4                                                                767740         680518         -11.36%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-4                                                               850676         763378         -10.26%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-4                                                              2518193        2312135        -8.18%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-4                                                             16866493       15626329       -7.35%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-4                                                                1757881        1628398        -7.37%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-4                                                                17663023       16271589       -7.88%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-4                                                            176791209      167930543      -5.01%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-4                                                  1063990        905986         -14.85%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-4                                                 1552706        1491273        -3.96%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-4                                                7085431        7204762        +1.68%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-4                                               61933174       65780593       +6.21%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-4                                                  7894814        7235089        -8.36%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-4                                                 13234773       12788289       -3.37%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-4                                                68093442       69590476       +2.20%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-4                                               618915485      618714002      -0.03%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-4                                              76510891       71662853       -6.34%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-4                                             131000752      126201633      -3.66%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-4                                            682868528      675057878      -1.14%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-4                                           6175974854     6163384524     -0.20%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-4                                                                 786677         697475         -11.34%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-4                                                                901140         823530         -8.61%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-4                                                               2198667        2090834        -4.90%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-4                                                              15027356       14562326       -3.09%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-4                                                                 6713828        6039313        -10.05%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-4                                                                8024832        7257989        -9.56%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-4                                                               20749802       20713857       -0.17%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-4                                                              146725258      148036247      +0.89%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-4                                                             70492000       60882484       -13.63%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-4                                                            80649745       75537762       -6.34%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-4                                                           204409961      200998441      -1.67%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-4                                                          1485385132     1430361710     -3.70%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-4                                                                    772884         702164         -9.15%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-4                                                                   878062         807307         -8.06%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-4                                                                  1807130        1769447        -2.09%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-4                                                                 11319035       11566698       +2.19%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-4                                                                    6634299        5886025        -11.28%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-4                                                                   7636080        6883710        -9.85%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-4                                                                  17109581       16688852       -2.46%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-4                                                                 115557933      114689321      -0.75%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-4                                                                68881042       58882513       -14.52%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-4                                                               75608469       69002832       -8.74%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-4                                                              172455230      166723770      -3.32%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-4                                                             1116699262     1130067522     +1.20%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1-4                                                        765051         689753         -9.84%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=10-4                                                       792220         699104         -11.75%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=100-4                                                      974001         938348         -3.66%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-4                                                     2941349        2900340        -1.39%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-4                                                        6492439        5790869        -10.81%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-4                                                       6686766        6298659        -5.80%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-4                                                      8694462        8021394        -7.74%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-4                                                     29239262       28524790       -2.44%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-4                                                    64543960       59829165       -7.30%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-4                                                   67137306       59977760       -10.66%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-4                                                  86524211       80345788       -7.14%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-4                                                 300418993      277779066      -7.54%
BenchmarkRangeQuery/expr=-a_one,steps=1-4                                                                             16456          16735          +1.70%
BenchmarkRangeQuery/expr=-a_one,steps=10-4                                                                            16695          16050          -3.86%
BenchmarkRangeQuery/expr=-a_one,steps=100-4                                                                           24839          23202          -6.59%
BenchmarkRangeQuery/expr=-a_one,steps=1000-4                                                                          89277          78638          -11.92%
BenchmarkRangeQuery/expr=-a_ten,steps=1-4                                                                             78491          69795          -11.08%
BenchmarkRangeQuery/expr=-a_ten,steps=10-4                                                                            78819          71975          -8.68%
BenchmarkRangeQuery/expr=-a_ten,steps=100-4                                                                           159161         140882         -11.48%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-4                                                                          801945         694778         -13.36%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-4                                                                         706523         620035         -12.24%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-4                                                                        720904         647710         -10.15%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-4                                                                       1542401        1371822        -11.06%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-4                                                                      8060421        7002051        -13.13%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-4                                                                      35447          28724          -18.97%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-4                                                                     36085          34845          -3.44%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-4                                                                    116549         108541         -6.87%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-4                                                                   827950         821716         -0.75%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-4                                                                      176675         155699         -11.87%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-4                                                                     230650         220550         -4.38%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-4                                                                    956533         925933         -3.20%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-4                                                                   7899870        7653202        -3.12%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-4                                                              1635885        1510476        -7.67%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-4                                                             2349669        2179525        -7.24%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-4                                                            10513118       10200682       -2.97%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-4                                                           87691862       86684269       -1.15%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-4                                                                  8076090        8140884        +0.80%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-4                                                                  79371745       78038306       -1.68%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-4                                                          943532697      895252727      -5.12%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-4                                                     72105          66640          -7.58%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-4                                                    75355          70512          -6.43%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-4                                                   106632         105914         -0.67%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-4                                                  448776         426228         -5.02%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-4                                                     175281         163049         -6.98%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-4                                                    203009         193321         -4.77%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-4                                                   573151         544609         -4.98%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-4                                                  4025428        3859601        -4.12%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-4                                             1249123        1093012        -12.50%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-4                                            1542008        1440204        -6.60%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-4                                           5964734        5595421        -6.19%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-4                                          46254448       44385735       -4.04%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-4                                                      67085          66567          -0.77%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-4                                                     74959          71288          -4.90%
…