@codesome 2020-06-16
Discussed offline: this [additional license in `bstream_test.go`](https://github.com/prometheus/prometheus/blob/f42ed03dc5be7290a913fe427cdf96f6d1bee792/tsdb/chunkenc/bstream_test.go#L14-L40) was added by mistake. It will be removed.